Started refactoring: new files for the models
