+++ billie/Model/RecognizedContentModel.swift
@@ -1,0 +1,12 @@
+//
+//  RecognizedContentModel.swift
+//  billie
+//
+//  Created by Pedro Muniz on 04/10/22.
+//
+
+import Foundation
+
+class RecognizedContent: ObservableObject {
+    @Published var items = [TextItem]()
+}
