+++ billie/Model/TextItemModel.swift
@@ -1,0 +1,18 @@
+//
+//  ScanData.swift
+//  billie
+//
+//  Created by Otávio Albuquerque on 08/09/22.
+//
+
+import Foundation
+import SwiftUI
+
+class TextItem: Identifiable {
+    var id: String
+    var text: String = ""
+    
+    init() {
+        id = UUID().uuidString
+    }
+}
