+++ billie/View/ActionButton.swift
@@ -88,4 +88,3 @@ struct manualEnterButton: ButtonStyle {
         }
     }
 }
-
